make note gating better, extract out code to reuse notes from framework

+++ Source/MPEWaveguideVoice.h
@@ -42,7 +42,7 @@ public:
     
     int compareTo(MPEWaveguideVoice* other);
     
-    float vol;
+    double vol;
 
     
 private:
